+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## ***CHANGELOG***
 
+### ***1.1.17***
+> - Store user settings of appearance and scaling for future
+    sessions.
+
 ### ***1.1.16***
 > - Disable 'submit' and 'attach file' button during the request proccessing.
 > - Added popup messages to confirm actions and display errors.
@@ -11,8 +15,9 @@
 
 ### ***1.0.12***
 > - Changed model details button to model info which allows the user to view
-   details, modelfile, parameter, template, system, adapter and license of the
-   loaded model.
+    details, modelfile, parameter, template, system, adapter and license of the
+    loaded model.
+
 > - Removed the model details functionality.
 
 ### ***1.0.0***